feat(hero): refine HeroLayout with pixel-perfect, animations and new shared folder
