+++ src/sections/home/frequentlyAskedQuestions/FrequentlyAskedQuestions.tsx
@@ -1,6 +1,6 @@
-import SectionWrapper from '@/components/layout/sectionWrapper/SectionWrapper'
 import AccordionSection from '@/sections/home/frequentlyAskedQuestions/components/AccordionSection'
 import ContactSection from '@/sections/home/frequentlyAskedQuestions/components/ContactSection'
+import SectionWrapper from '@/sections/shared/sectionWrapper/SectionWrapper'
 
 export const FrequentlyAskedQuestions = () => {
   return (
