+++ src/sections/home/impact/Impact.tsx
@@ -1,5 +1,5 @@
-import SectionWrapper from '@/components/layout/sectionWrapper/SectionWrapper'
 import { Card } from '@/components/ui/Card'
+import SectionWrapper from '@/sections/shared/sectionWrapper/SectionWrapper'
 
 const impactProps = [
   {
