+++ src/sections/home/valueProposition/ValueProposition.tsx
@@ -1,5 +1,5 @@
-import SectionWrapper from "@/components/layout/sectionWrapper/SectionWrapper";
 import { Card } from "@/components/ui/Card";
+import SectionWrapper from "@/sections/shared/sectionWrapper/SectionWrapper";
 import { Gift, Globe2, HeartHandshake, MapPin, Rocket, Users } from "lucide-react";
 
 const valueProps = [
